Working token bucket layer

+++ src/rate_limit/rate.rs
@@ -3,7 +3,7 @@ use std::time::Duration;
 /// A rate of requests per time period.
 #[derive(Debug, Copy, Clone)]
 pub struct Rate {
-    num: u64,
+    num: usize,
     per: Duration,
 }
 
@@ -13,14 +13,14 @@ impl Rate {
     /// # Panics
     ///
     /// This function panics if `num` or `per` is 0.
-    pub fn new(num: u64, per: Duration) -> Self {
+    pub fn new(num: usize, per: Duration) -> Self {
         assert!(num > 0);
         assert!(per > Duration::from_millis(0));
 
         Rate { num, per }
     }
 
-    pub(crate) fn num(&self) -> u64 {
+    pub(crate) fn num(&self) -> usize {
         self.num
     }
